feat: adds primitives crate
