+++ crates/sleppa_configuration/Cargo.toml
@@ -25,10 +25,10 @@ rust-version.workspace = true
 [dependencies]
 
 # External dependencies
-serde = { version = "^1.0.160", features = ["derive"] }
-toml = { version = "^0.7.3" }
-regex = { version = "^1.8.1" }
-tempfile = { version = "^3.5.0" }
+regex = { version = "^1.8" }
+serde = { version = "^1.0", features = ["derive"] }
+tempfile = { version = "^3.5" }
+toml = { version = "^0.7" }
 
 # Errors and logs processing
 thiserror = { workspace = true }
