+++ crates/sleppa_primitives/src/repositories/errors.rs
